teacher_portal/requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
packaging==25.0
psycopg2-binary==2.9.11
py-dotenv==0.1
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
